Creation of the transition probabilities logic, also created alot of helper functions
The transition probability 3d array contains all the probabilities of taking an course from a state and then passing or failing the course. The array could be improved since it contains alot of 0 probabilites values for impossible actions.

TODO:
Max skill level is now set to 4,4,4 the user should be able to give this as input (the goal they want to achieve)

When a student fails it stays in the same skill state, the actual updating formula still needs to be implemented for passing and failing.